pyproject.toml
[project]
name = "gleam"
version = "0.1.0"
description = "Neural Phong shader (cGAN) with ModernGL renderer — SIGK Projekt 3"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "moderngl>=5.12",
    "pyrr>=0.10.3",
    "pywavefront>=1.3.3",
    "pillow>=10.3",
    "numpy>=1.26,<2.3",
    "torch>=2.4",
    "torchvision>=0.19",
    "lpips>=0.1.4",
    "pytorch-msssim>=1.0",
    "scikit-image>=0.24",
    "scipy>=1.13",
    "opencv-python>=4.10",
    "flip-evaluator>=1.6",
    "typer>=0.12",
    "loguru>=0.7.2",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "pandas>=2.2",
]

[project.scripts]
gleam = "gleam.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gleam"]

# CUDA wheels for Blackwell (DGX Spark) — Linux only. macOS keeps the default
# PyPI wheels with MPS support.
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux'" },
]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]
